Cleaning up duplicate vertices and triangles
In my postprocessing the duplicate nodes caused difficulties (compared to the Triangle Mesher which has unique vertices and triangles). These few lines clean the data.

In my experience, the 'trick' with the donut as you show in the examples is not even necessary to add. I obtain valid meshes with overlapping polygons.