Catch gmsh not running in main thread. Fixes #13
I've tried adding a test for this; it's easy to reproduce in a standalone script by starting a thread which calls the import.
It's not easy to test for with pytest, since pytest runs the tests in a single session. In that case, pandamesh is already imported, and the threaded import statement does nothing.